pyproject.toml
[tool.poetry]
name = "socket"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = "^5.7.2"
uvicorn = "^0.19.0"
redis = "^4.3.4"
dependency-injector = "^4.40.0"
arq = "^0.24.0"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
